moved Home.js to pages and cleaned up propertyCard.js code
